requirements.txt
openai>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
openai-whisper>=20231117
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10
PyGetWindow>=0.0.9
pyautogui>=0.9.54
transformers>=4.35.0
llama-cpp-python>=0.2.0
sounddevice>=0.4.6
numpy<2.0.0
pyperclip>=1.8.2
google-generativeai>=0.3.0 